requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
fisx==1.3.1
fonttools==4.47.2
guidata==3.1.0
guiqwt==4.4.4
h5py==3.10.0
idna==3.6
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
Pint==0.23
ply==3.11
psutil==5.9.7
pyepics==3.5.2
PyMca5==5.9.2
pyparsing==3.1.1
pypylon==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
pytango==9.5.0
python-dateutil==2.8.2
PythonQwt==0.11.2
QtPy==2.4.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
taurus==5.1.8
taurus-pyqtgraph==0.8.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
